lazy_rc

Struct LazyRc

pub struct LazyRc<T> { /* private fields */ }

A single-threaded reference-counting pointer, akin to Rc<T>, but with lazy initialization

Implementations

impl<T> LazyRc<T>

pub const fn empty() -> Self

Create a new LazyRc<T> that is initially empty and that contains no default initializer.

The “inner” value will be initialized on first access. Default initialization is not supported by this instance!

pub fn with_default_init<U>(default_init: U) -> Self
where U: Fn() -> T + Sync + 'static,

Create a new LazyRc<T> that is initially empty and that contains the given default initializer.

The “inner” value will be initialized on first access. Default initialization is supported by this instance.

pub fn with_failable_default_init<U>(default_init: U) -> Self
where U: Fn() -> IoResult<T> + Sync + 'static,

Create a new LazyRc<T> that is initially empty and that contains the given failable default initializer.

The “inner” value will be initialized on first access. Default initialization is supported by this instance.

pub fn is_initialized(&self) -> bool

Returns true, if and only if th “inner” value is initialized.

pub fn or_init(&self) -> Rc<T>

Returns a pointer to the existing “inner” value, or tries to initialize the value right now.

If and only if the “inner” value is not initialized yet, the “inner” value is set to the return value of the default initializer and a new Rc<T> pointer to the “inner” value is returned. The default initializer must be infailable, otherwise use or_try_init()!

Warning: This function panics, if no default initializer is available, or of the default initializer is failable!

pub fn or_try_init(&self) -> Result<Rc<T>, InitError>

Returns a pointer to the existing “inner” value, or tries to initialize the value right now.

If and only if the “inner” value is not initialized yet, the “inner” value is set to the return value of the default initializer and a new Rc<T> pointer to the “inner” value is returned. If the default initializer fails, the error is passed through.

If no default initializer is available, an error of type NoDefaultInitializer is returned.

pub fn or_init_with<F>(&self, init_fn: F) -> Rc<T>
where F: FnOnce() -> T,

Returns a pointer to the existing “inner” value, or initializes the value right now.

If and only if the “inner” value is not initialized yet, the function init_fn() is called to create the value. The “inner” value is then set to the return value of init_fn() and a new Rc<T> pointer to the “inner” value is returned.

pub fn or_try_init_with<E, F>(&self, init_fn: F) -> Result<Rc<T>, E>
where F: FnOnce() -> Result<T, E>,

Returns a pointer to the existing “inner” value, or tries to initializes the value right now.

If and only if the “inner” value is not initialized yet, the function init_fn() is called to create the value. In case that init_fn() returns an error, that error is passed through and the “inner” value remains in the uninitialized state for now. If the “inner” value already existed or if it was created successfully just now, a new Rc<T> pointer to the “inner” value is returned.

pub fn unwrap(&self) -> Rc<T>

An alias for the or_init() function.

pub fn map<U, F>(&self, map_fn: F) -> Option<U>
where F: FnOnce(&Rc<T>) -> U,

Applies function map_fn() to the “inner”, if already initialized.

If and only if the “inner” value already is initialize, the function map_fn() is called with a reference to the “inner” value and its return value is passed through. Otherwise the function map_fn() is not called and None is returned.

pub fn value(&self) -> Option<Rc<T>>

Returns a pointer to the “inner” value, if already initialized.

If and only if the “inner” value already is initialized, the function returns a new Rc<T> pointer to the “inner” value. Otherwise, if the “inner” value is not initialized yet, the value remains in the uninitialized state and the function returns None.

pub fn take(&mut self) -> Option<Rc<T>>

Takes the “inner” value out of this LazyRc<T> instance, if already initialized.

If and only if the “inner” value already is initialized, the function returns the Rc<T> pointer to the current “inner” value and resets this LazyRc<T> instance’ “inner” value to the uninitialized state. Otherwise, the function simply returns None.